Add the DOCKER_COMPOSER_BIN to docker-compose.js for E2E tests.
A environment variable named DOCKER_COMPOSER_BIN can now be used
to specify the name of the docker-compose executable in the system,
defaults to "docker-compose". This is useful when using WSL, where
docker-compose is v2 (which won't work for the Woo E2E tests)
and the v1 executable is "docker-compose-v1".